+++ back/db/seeds/citizenlab.rb
@@ -32,7 +32,9 @@ AppConfiguration.create!(
     customizable_homepage_banner: {
       allowed: true,
       enabled: true,
-      layout: 'full_width_banner_layout'
+      layout: 'full_width_banner_layout',
+      cta_signed_out_type: 'sign_up_button',
+      cta_signed_in_type: 'no_button'
     },
     password_login: {
       enabled: true,