+++ front/app/i18n/nl-BE.ts
@@ -6,9 +6,10 @@ import { formatTranslationMessages } from './';
 addLocaleData(nlLocaleData);
 
 const nlBETranslationMessages = require('translations/nl-BE.json');
-const translationMessages = formatTranslationMessages(
-  'nl-BE',
-  nlBETranslationMessages
-);
+const nlBEAdminTranslationMessages = require('translations/admin/nl-BE.json');
+const translationMessages = formatTranslationMessages('nl-BE', {
+  ...nlBETranslationMessages,
+  ...nlBEAdminTranslationMessages,
+});
 
 export default translationMessages;