+++ front/app/services/__mocks__/appConfiguration.ts
@@ -40,6 +40,16 @@ export const getAppConfigurationData = (
         allowed: true,
         enabled: true,
         layout: 'full_width_banner_layout',
+        cta_signed_out_type: 'sign_up_button',
+        cta_signed_out_customized_button: {
+          text: { en: 'Click' },
+          url: 'https://www.wonder.ville/promo',
+        },
+        cta_signed_in_type: 'no_button',
+        cta_signed_in_customized_button: {
+          text: { en: 'Click' },
+          url: 'https://www.wonder.ville/promo2',
+        },
       },
       participatory_budgeting: {
         allowed: true,